Credit Exit & Input Text Optimisations
Made it so you can exit the credits through use of input and implemented the exiting to menu through scrolling till the end.
Changed the Hover Text to get component on the player since it's already used for tracking, this is instead of searching the hierarchy for a player input component which is MUCH faster.